BuildLangModel: mark cached files against the 'visited_pages' when processed so as not to refetch it when we end up grabbing a few more pages off wikipedia after all.
For this to work, we also need to cache the collected *titles* (links!) as we would otherwise loose those when we'ld load the cached file (content) rather than the wikipedia itself! Newly discovered links (titles) are appended to the cache file (`links.txt`) to keep the initial processing order intact: the cached set is *appended* to the `start_pages` when the process starts.